requirements.txt
Cython
SQLAlchemy
pandas==0.25.1
dask
beautifulsoup4
python_dateutil
python-dotenv
